Refactor file structures (actions and its types)
